Refactor privacy-pool-node-prover to WithdrawalProofGenerator
- Rename privacy-pool-node-prover.ts to WithdrawalProofGenerator.ts
- Rename class from PrivacyPoolNodeProver to WithdrawalProofGenerator
- Improve code organization with better TypeScript interfaces
- Add configurable circuit paths with automatic discovery
- Enhanced error handling and logging
- Remove hardcoded paths for better portability
- Clean up example usage code to keep module focused
- Update import in simple-e2e-real-proofs.ts
- Verified functionality: proof generation and verification working correctly
